+++ app/src/main/java/com/sourcepoint/test_project/MainActivity.java
@@ -3,61 +3,79 @@ package com.sourcepoint.test_project;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
+import android.view.ViewGroup;
+import android.webkit.WebView;
 
-import com.sourcepoint.cmplibrary.ConsentLib;
-import com.sourcepoint.cmplibrary.ConsentLibException;
-import com.sourcepoint.cmplibrary.CustomVendorConsent;
 
-import java.util.HashSet;
+import com.sourcepoint.gdpr_cmplibrary.GDPRConsentLib;
+import com.sourcepoint.gdpr_cmplibrary.GDPRUserConsent;
 
 public class MainActivity extends AppCompatActivity {
     private static final String TAG = "MainActivity";
 
-    private ConsentLib consentLib;
+    private GDPRConsentLib gdprConsentLib;
+    private ViewGroup mainViewGroup;
 
-    private ConsentLib buildAndRunConsentLib(Boolean showPM) throws ConsentLibException {
-        return ConsentLib.newBuilder(22, "mobile.demo", 2372,"5c0e81b7d74b3c30c6852301",this)
-                .setStage(true)
-                .setViewGroup(findViewById(android.R.id.content))
-                .setShowPM(showPM)
-                .setOnMessageReady(consentLib -> Log.i(TAG, "onMessageReady"))
-                .setOnConsentReady(consentLib -> consentLib.getCustomVendorConsents(results -> {
-                    HashSet<CustomVendorConsent> consents = (HashSet) results;
-                    for(CustomVendorConsent consent : consents)
-                        Log.i(TAG, "Consented to: "+consent);
-                }))
-                .setOnErrorOccurred(c -> Log.i(TAG, "Something went wrong: ", c.error))
+    private void showMessageWebView(WebView webView) {
+        webView.setLayoutParams(new ViewGroup.LayoutParams(0, 0));
+        webView.getLayoutParams().height = ViewGroup.LayoutParams.MATCH_PARENT;
+        webView.getLayoutParams().width = ViewGroup.LayoutParams.MATCH_PARENT;
+        webView.bringToFront();
+        webView.requestLayout();
+        mainViewGroup.addView(webView);
+    }
+    private void removeWebView(WebView webView) {
+        if(webView.getParent() != null)
+            mainViewGroup.removeView(webView);
+
+    }
+
+    private GDPRConsentLib buildGDPRConsentLib() {
+        return GDPRConsentLib.newBuilder(22, "mobile.demo", 2372,"5c0e81b7d74b3c30c6852301",this)
+                .setStagingCampaign(true)
+                .setAuthId("gdpr-test")
+                .setOnConsentUIReady(consentLib -> {
+                    showMessageWebView(consentLib.webView);
+                    Log.i(TAG, "onConsentUIReady");
+                })
+                .setOnConsentUIFinished(consentLib -> {
+                    removeWebView(consentLib.webView);
+                    Log.i(TAG, "onConsentUIFinished");
+                })
+                .setOnConsentReady(consentLib -> {
+                    Log.i(TAG, "onConsentReady");
+                    GDPRUserConsent consent = consentLib.userConsent;
+                    for (String vendorId : consent.acceptedVendors) {
+                        Log.i(TAG, "The vendor " + vendorId + " was accepted.");
+                    }
+                    for (String purposeId : consent.acceptedCategories) {
+                        Log.i(TAG, "The category " + purposeId + " was accepted.");
+                    }
+                })
+                .setOnError(consentLib -> {
+                    Log.e(TAG, "Something went wrong: ", consentLib.error);
+                    Log.i(TAG, "ConsentLibErrorMessage: " + consentLib.error.consentLibErrorMessage);
+                    removeWebView(consentLib.webView);
+                })
                 .build();
     }
 
     @Override
     protected void onResume() {
         super.onResume();
-        try {
-            consentLib = buildAndRunConsentLib(false);
-            consentLib.run();
-        } catch (ConsentLibException e) {
-            e.printStackTrace();
-        }
+        gdprConsentLib = buildGDPRConsentLib();
+        gdprConsentLib.run();
     }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+        mainViewGroup = findViewById(android.R.id.content);
         findViewById(R.id.review_consents).setOnClickListener(_v -> {
-            try {
-                consentLib = buildAndRunConsentLib(true);
-                consentLib.run();
-            } catch (ConsentLibException e) {
-                e.printStackTrace();
-            }
-        });
-    }
+            gdprConsentLib = buildGDPRConsentLib();
+            gdprConsentLib.showPm();
 
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-        if(consentLib != null ) { consentLib.destroy(); }
+        });
     }
 }