+++ cmplibrary/src/main/java/com/sourcepoint/cmplibrary/CustomVendorConsent.java
@@ -1,7 +1,0 @@
-package com.sourcepoint.cmplibrary;
-
-public class CustomVendorConsent extends Consent {
-    CustomVendorConsent(String id, String name) {
-        super(id, name, "CustomVendorConsent");
-    }
-}