+++ cmplibrary/src/main/java/com/sourcepoint/gdpr_cmplibrary/Consent.java
@@ -1,4 +1,4 @@
-package com.sourcepoint.cmplibrary;
+package com.sourcepoint.gdpr_cmplibrary;
 
 import org.json.JSONException;
 import org.json.JSONObject;