+++ cmplibrary/src/main/java/com/sourcepoint/gdpr_cmplibrary/ConsentWebView.java
@@ -1,4 +1,4 @@
-package com.sourcepoint.cmplibrary;
+package com.sourcepoint.gdpr_cmplibrary;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
@@ -17,7 +17,6 @@ import android.util.Log;
 import android.view.KeyEvent;
 import android.view.ViewGroup;
 import android.webkit.CookieManager;
-import android.webkit.CookieSyncManager;
 import android.webkit.JavascriptInterface;
 import android.webkit.RenderProcessGoneDetail;
 import android.webkit.SslErrorHandler;
@@ -27,75 +26,116 @@ import android.webkit.WebResourceRequest;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 
+import org.json.JSONException;
+import org.json.JSONObject;
+
 import java.util.HashSet;
 
 abstract public class ConsentWebView extends WebView {
     private static final String TAG = "ConsentWebView";
 
+
+    // TODO: pass this script to a .js file and return it as a string in this method
+    private String getJSInjection(){
+        return "addEventListener('message', handleEvent);\n" +
+                "function handleEvent(event) {\n" +
+                "    try {\n" +
+                "        JSReceiver.log(JSON.stringify(event.data, null, 2));\n" +
+                "        if (event.data.name === 'sp.showMessage') {\n" +
+                "            JSReceiver.onConsentUIReady();\n" +
+                "            return;\n" +
+                "        }\n" +
+                "        const data = eventData(event);\n" +
+                "        JSReceiver.log(JSON.stringify(data, null, 2));\n" +
+                "        if(data.type) {\n" +
+                "            if(data.type === 1) JSReceiver.onSavePM(JSON.stringify(data.payload));\n" +
+                "            else JSReceiver.onAction(data.type);\n" +
+                "        }\n" +
+                "    } catch (err) {\n" +
+                "        JSReceiver.log(err.stack);\n" +
+                "    };\n" +
+                "};\n" +
+                "\n" +
+                "function eventData(event) {\n" +
+                "    return isFromPM(event) ? dataFromPM(event) : dataFromMessage(event);\n" +
+                "};\n" +
+                "\n" +
+                "function isFromPM(event) {\n" +
+                "    return !!event.data.payload;\n" +
+                "};\n" +
+                "\n" +
+                "function dataFromMessage(msgEvent) {\n" +
+                "    return {\n" +
+                "        name: msgEvent.data.name,\n" +
+                "        type: msgEvent.data.actions.length ? msgEvent.data.actions[0].data.type : null\n" +
+                "    };\n" +
+                "};\n" +
+                "\n" +
+                "function dataFromPM(pmEvent) {\n" +
+                "    const data = {\n" +
+                "        name: pmEvent.data.name,\n" +
+                "        type: pmEvent.data ? pmEvent.data.payload.actionType : null,\n" +
+                "    };\n" +
+                "    if(data.type === 1) data.payload = userConsents(pmEvent.data.payload);\n" +
+                "    return data;\n" +
+                "};\n" +
+                "\n" +
+                "function userConsents(payload){\n" +
+                "    return {\n" +
+                "        acceptedVendors: payload.consents.vendors.accepted,\n" +
+                "        acceptedCategories: payload.consents.categories.accepted\n" +
+                "    }\n" +
+                "}";
+    }
+
     @SuppressWarnings("unused")
     private class MessageInterface {
 
+        @JavascriptInterface
+        public void log(String tag, String msg){
+            Log.i(tag, msg);
+        }
+
+        @JavascriptInterface
+        public void log(String msg){
+            Log.i("JS", msg);
+        }
+
         // called when message is about to be shown
         @JavascriptInterface
-        public void onMessageReady() {
-            Log.d("onMessageReady", "called");
-            ConsentWebView.this.flushOrSyncCookies();
+        public void onConsentUIReady() {
+            Log.d("onConsentUIReady", "called");
             ConsentWebView.this.onMessageReady();
         }
 
         // called when a choice is selected on the message
         @JavascriptInterface
-        public void onMessageChoiceSelect(int choiceId, int choiceType) {
-            Log.d("onMessageChoiceSelect", "called");
+        public void onAction(int choiceType) {
+            Log.d("onAction", "called");
             if (ConsentWebView.this.hasLostInternetConnection()) {
-                ConsentWebView.this.onErrorOccurred(new ConsentLibException.NoInternetConnectionException());
+                ConsentWebView.this.onError(new ConsentLibException.NoInternetConnectionException());
             }
-            ConsentWebView.this.onMessageChoiceSelect(choiceId, choiceType);
+            ConsentWebView.this.onAction(choiceType);
         }
 
-        //called when user takes action on privacy manager
+        // called when a choice is selected on the message
         @JavascriptInterface
-        public void onPrivacyManagerAction(String pmData) {
-            Log.d("onPrivacyManagerAction", "called");
-        }
-
-        @JavascriptInterface
-        public void onMessageChoiceError(String errorType) {
-            onErrorOccurred(errorType);
-            Log.d("onMessageChoiceError", "called");
-        }
-
-        // called when interaction with message is complete
-        @JavascriptInterface
-        public void onConsentReady(String consentUUID, String euConsent) {
-            Log.d("onConsentReady", "called");
-            ConsentWebView.this.flushOrSyncCookies();
-            ConsentWebView.this.onConsentReady(euConsent, consentUUID);
-        }
-
-        //called when privacy manager cancel button is tapped
-        @JavascriptInterface
-        public void onPMCancel() {
-            Log.d("onPMCancel", "called");
-        }
-
-        // called when message or pm need to shown
-        @JavascriptInterface
-        public void onSPPMObjectReady() {
-            Log.d("onSPPMObjectReady", "called");
-            if (isShowPM) {
-                ConsentWebView.this.flushOrSyncCookies();
-                ConsentWebView.this.onMessageReady();
+        public void onSavePM(String payloadStr) throws JSONException, ConsentLibException {
+            Log.d("onSavePM", "called");
+            if (ConsentWebView.this.hasLostInternetConnection()) {
+                ConsentWebView.this.onError(new ConsentLibException.NoInternetConnectionException());
             }
+            JSONObject payloadJson = new JSONObject(payloadStr);
+            ConsentWebView.this.onSavePM(new GDPRUserConsent(payloadJson));
         }
 
         //called when an error is occured while loading web-view
         @JavascriptInterface
-        public void onErrorOccurred(String errorType) {
+        public void onError(String errorType) {
             ConsentLibException error = ConsentWebView.this.hasLostInternetConnection() ?
                     new ConsentLibException.NoInternetConnectionException() :
                     new ConsentLibException("Something went wrong in the javascript world.");
-            ConsentWebView.this.onErrorOccurred(error);
+            ConsentWebView.this.onError(error);
         }
         // xhr logger
         @JavascriptInterface
@@ -131,15 +171,13 @@ abstract public class ConsentWebView extends WebView {
 
     private long timeoutMillisec;
     private ConnectionPool connectionPool;
-    private boolean isShowPM = false;
 
     public static long DEFAULT_TIMEOUT = 10000;
 
-    public ConsentWebView(Context context, long timeoutMillisec, boolean isShowPM) {
+    public ConsentWebView(Context context, long timeoutMillisec) {
         super(context);
         this.timeoutMillisec = timeoutMillisec;
         connectionPool = new ConnectionPool();
-        this.isShowPM = isShowPM;
         setup();
     }
 
@@ -172,9 +210,8 @@ abstract public class ConsentWebView extends WebView {
                 setWebContentsDebuggingEnabled(true);
                 enableSlowWholeDocumentDraw();
             }
-            CookieManager.getInstance().setAcceptThirdPartyCookies(this, true);
         }
-        CookieManager.getInstance().setAcceptCookie(true);
+        CookieManager.getInstance().setAcceptCookie(false);
         getSettings().setAppCacheEnabled(false);
         getSettings().setBuiltInZoomControls(false);
         getSettings().setSupportZoom(false);
@@ -190,7 +227,7 @@ abstract public class ConsentWebView extends WebView {
                 connectionPool.add(url);
                 Runnable run = () -> {
                     if (connectionPool.contains(url))
-                        onErrorOccurred(new ConsentLibException.ApiException("TIMED OUT: " + url));
+                        onError(new ConsentLibException.ApiException("TIMED OUT: " + url));
                 };
                 Handler myHandler = new Handler(Looper.myLooper());
                 myHandler.postDelayed(run, timeoutMillisec);
@@ -199,36 +236,37 @@ abstract public class ConsentWebView extends WebView {
             @Override
             public void onPageFinished(WebView view, String url) {
                 super.onPageFinished(view, url);
-                flushOrSyncCookies();
                 connectionPool.remove(url);
+                //view.loadUrl("javascript:" + "addEventListener('message', SDK.onEvent('oie'))");
+                view.loadUrl("javascript:" + getJSInjection());
             }
 
             @Override
             public void onReceivedError(WebView view, WebResourceRequest request, WebResourceError error) {
                 super.onReceivedError(view, request, error);
                 Log.d(TAG, "onReceivedError: " + error.toString());
-                onErrorOccurred(new ConsentLibException.ApiException(error.toString()));
+                onError(new ConsentLibException.ApiException(error.toString()));
             }
 
             @Override
             public void onReceivedError(WebView view, int errorCode, String description, String failingUrl) {
                 super.onReceivedError(view, errorCode, description, failingUrl);
                 Log.d(TAG, "onReceivedError: Error " + errorCode + ": " + description);
-                onErrorOccurred(new ConsentLibException.ApiException(description));
+                onError(new ConsentLibException.ApiException(description));
             }
 
             @Override
             public void onReceivedSslError(WebView view, SslErrorHandler handler, SslError error) {
                 super.onReceivedSslError(view, handler, error);
                 Log.d(TAG, "onReceivedSslError: Error " + error);
-                onErrorOccurred(new ConsentLibException.ApiException(error.toString()));
+                onError(new ConsentLibException.ApiException(error.toString()));
             }
 
             @Override
             public boolean onRenderProcessGone(WebView view, RenderProcessGoneDetail detail) {
                 String message = "The WebView rendering process crashed!";
                 Log.e(TAG, message);
-                onErrorOccurred(new ConsentLibException(message));
+                onError(new ConsentLibException(message));
                 return false;
             }
         });
@@ -265,39 +303,28 @@ abstract public class ConsentWebView extends WebView {
         return activeNetwork == null || !activeNetwork.isConnectedOrConnecting();
     }
 
-    private void flushOrSyncCookies() {
-        // forces the cookies sync between RAM and local storage
-        // https://developer.android.com/reference/android/webkit/CookieSyncManager
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP)
-            CookieManager.getInstance().flush();
-        else CookieSyncManager.getInstance().sync();
-    }
-
-    @Override
-    protected void onDetachedFromWindow() {
-        super.onDetachedFromWindow();
-        flushOrSyncCookies();
-    }
-
     abstract public void onMessageReady();
 
-    abstract public void onErrorOccurred(ConsentLibException error);
+    abstract public void onError(ConsentLibException error);
 
-    abstract public void onConsentReady(String euConsent, String consentUUID);
+    abstract public void onAction(int choiceType);
 
-    abstract public void onMessageChoiceSelect(int choiceType, int choiceId);
+    abstract public void onSavePM(GDPRUserConsent GDPRUserConsent);
 
-    public void loadMessage(String messageUrl) throws ConsentLibException.NoInternetConnectionException {
+
+    public void loadConsentMsgFromUrl(String url) throws ConsentLibException.NoInternetConnectionException {
         if (hasLostInternetConnection())
             throw new ConsentLibException.NoInternetConnectionException();
 
         // On API level >= 21, the JavascriptInterface is not injected on the page until the *second* page load
         // so we need to issue blank load with loadData
         loadData("", "text/html", null);
-        Log.d(TAG, "Loading Webview with: " + messageUrl);
+        Log.d(TAG, "Loading Webview with: " + url);
         Log.d(TAG, "User-Agent: " + getSettings().getUserAgentString());
-        loadUrl(messageUrl);
+        loadUrl(url);
     }
+
+
 
     public void display() {
         setLayoutParams(new ViewGroup.LayoutParams(0, 0));